// .codex-backups/revoke-invites/types/api.ts
export interface Branch {
  id: string;
  organizationId: string;
  name: string;
  managerId: string | null;
  createdBy: string;
}